+++ db/migrate/20230116173500_add_approved_at_to_reports.rb
@@ -1,0 +1,5 @@
+class AddApprovedAtToReports < ActiveRecord::Migration[6.1]
+  def change
+    add_column :reports, :approved_at, :datetime
+  end
+end